backend/pyproject.toml
[project]
dependencies = [
  "streamlit>=1.48.1",
  "youtube-transcript-api>=1.2.2",
  "yt-dlp>=2025.8.27",
  "google-cloud-translate>=3.15.0",
  "langdetect>=1.0.9",
  "google-generativeai>=0.4.0",
  "chromadb>=0.4.24",
  "langchain>=0.1.12",
  "langchain-google-genai>=0.0.11",
  "fastapi>=0.110.0",
  "pydantic[email]>=2.6.4",
  "pymongo>=4.6.2",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "pyjwt>=2.10.1",
  "motor>=3.7.1",
  "langchain-community>=0.3.29",
  "langchain-chroma>=0.2.6",
  "langchain-text-splitters>=0.3.11",
  "langchain-mongodb>=0.7.0",
]
description = "YouTube video summarizer"
name = "yt-summarizer-backend"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"
